User/Lib: deep audit + แก้ bug 25 จุด + อัปเดตเอกสารข้อจำกัด
## Bug Fixes (20 ไฟล์)
### CRITICAL — crash / ข้อมูลเสียหาย
- ESC, ServoCluster: PWM duty ใช้ PWM_GetPeriod() แทน hardcode 65535
- TM1637: INT16_MIN overflow ใช้ unsigned arithmetic
- TM1650: WaitKey + key release timeout 5s/2s
- NeoPixel: FillGradient กัน LEDw â‰¤ 1 → DIV/0
- WS2815Matrix: ลด max 32x32→8x8 (RAM 2KB); temp_buffer stack→static
- WS2815Matrix: FadeIn/FadeOut กัน steps=0
- MAX7219: ลบ SPI_SimpleInit ภายใน → ผู้ใช้ต้องเรียกเอง
- OLED: union buffer 3 ขนาด (RAM 1920→1024B)
- OLED: Font_8x16/12x16 .data=NULL → fallback Font_6x8
- SoftUART: baud=0 guard
- P10: bounds check buffer_size

### HIGH — silent failure / reliability
- PCA9685: เช็ค I2C return ทุก _read_reg/_write_reg
- AS5600: multi-byte I2C read (atomic, ไม่ torn read)
- nRF24L01: Delay_Ms(1) ใน TX busy-wait + FLUSH_TX บน timeout
- INA219: error sentinel -1.0f/-999f → NAN
- ESP01: HTTP URL length pre-check ก่อน snprintf
- HC05: Flush timeout 100ms
- GPS: _split_nmea guard max_fields==0
- DHT: timing calibration DHT_LOOP_OVERHEAD_US=3
- BH1750/SHT3x: !=0 → !=I2C_OK
- NTC10K: guard t0_kelvin≤0, b_value≤0
- MQGas: ADC_EnableChannel ใน Init

## เอกสาร (20 ไฟล์)
- LIMITATIONS.md: ไฟล์กลาง — ข้อจำกัดรวม 71 ไลบรารี, Package Compatibility, Timer Conflict Map
- 18 README.md: เพิ่ม ## ข้อจำกัด พร้อมลิงก์ไป LIMITATIONS.md

+++ User/Lib/TM1637/TM1637.c
@@ -255,7 +255,7 @@ void TM1637_Clear(TM1637_Handle* handle) {
 void TM1637_DisplayNumber(TM1637_Handle* handle, int16_t number, bool leading_zero) {
     bool negative = (number < 0);
     if (negative) {
-        number = -number;
+        number = -(int16_t)((uint16_t)number);
     }
     
     // Clear buffer
